Use best first search to get next square
Instead of placing the next value in the next free space, we now look
for the square with the least amount of possible values to put a value
in. This should improve the search and limit the size of the search
tree by a huge margin, improving perfomance for the cost of complexity.